web_client/requirements_web.txt
transformers==4.32.0
pyyaml==5.3.1
gdown==4.7.1
torch==1.11.0
streamlit==1.24.0
protobuf==3.20.*
zipfile36
